Updated Placement test file, and deleted the sitePlacement test file
